\name{acb_age_profile_data_2526}
\alias{acb_age_profile_data_2526}
\docType{data}
\title{
ACB age profile, 2025-2026, Valencia Basket
}
\description{
Roster of the ACB team Valencia Basket for the 2025-2026 season and its
players' age profiles.
}
\usage{acb_age_profile_data_2526}
\format{Data frame with 15 rows and 7 columns.}
\source{\url{https://acb.com/}}
\keyword{datasets}
