\name{acb_sticker_data_2526}
\alias{acb_sticker_data_2526}
\docType{data}
\title{
ACB sticker data, 2025-2026
}
\description{
This data frame contains an example of the usage and context statistics
for the Real Madrid player Alberto Abalde to illustrate the creation
of the sticker plot.
}
\usage{acb_sticker_data_2526}
\format{Data frame with 11 rows and 9 columns.}
\source{\url{https://acb.com/}}
\keyword{datasets}
