\name{acb_usage_act_data_2526}
\alias{acb_usage_act_data_2526}
\docType{data}
\title{
ACB usage action data, 2025-2026
}
\description{
Sample of possessions played and ended for a game from the ACB 2025-2026 season
and the particular number of actions used by the player to end the possession.
}
\usage{acb_usage_act_data_2526}
\format{Data frame with 75 rows and 8 columns.}
\source{\url{https://acb.com/}}
\keyword{datasets}
