\name{acb_usage_data_2526}
\alias{acb_usage_data_2526}
\docType{data}
\title{
ACB usage data, 2025-2026
}
\description{
Sample of possessions played and ended for a game from the ACB 2025-2026 season.
For each player we have the number of possessions he played and the ones he ended.
}
\usage{acb_usage_data_2526}
\format{Data frame with 62 rows and 9 columns.}
\source{\url{https://acb.com/}}
\keyword{datasets}
