% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capit_two_words.R
\name{capit_two_words}
\alias{capit_two_words}
\title{Capitalize two-word strings}
\usage{
capit_two_words(two_word_string)
}
\arguments{
\item{two_word_string}{Two-word string.}
}
\value{
Vector with the two words capitalized.
}
\description{
Ancillary function to capitalize the first letter of both words in a 
two-word string. This can be used for example to capitalize the teams 
names for the plots title.
}
\examples{
capit_two_words("valencia basket")

}
