% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_divide_court_zones.R
\name{do_divide_court_zones}
\alias{do_divide_court_zones}
\title{Zones of the basketball court}
\usage{
do_divide_court_zones(data_shots)
}
\arguments{
\item{data_shots}{Shooting data frame.}
}
\value{
The shooting data frame with a new column called "location" indicating 
the zone from which each shot was taken.
}
\description{
Divide the basketball court into 10 zones.
}
\examples{
\dontrun{
do_divide_court_zones(acb_shooting_data_2425)
}

}
\author{
Guillermo Vinue
}
