% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_join_games_bio.R
\name{do_join_games_bio}
\alias{do_join_games_bio}
\title{Join games and players' info}
\usage{
do_join_games_bio(competition, df_games, df_rosters)
}
\arguments{
\item{competition}{String. Options are "ACB", "Euroleague" and "Eurocup".}

\item{df_games}{Data frame with the games.}

\item{df_rosters}{Data frame with the biography of the roster players.}
}
\value{
Data frame.
}
\description{
This function calls the needed ancillary functions to join the games played 
by all the players in the desired competition (currently ACB, Euroleague
and Eurocup) with their personal details.
}
\examples{
df <- do_join_games_bio("ACB", acb_games_1718, acb_players_1718)
          
}
\seealso{
\code{\link{join_players_bio_age_acb}}, \code{\link{join_players_bio_age_euro}}
}
\author{
Guillermo Vinue
}
