% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_map_nats.R
\name{do_map_nats}
\alias{do_map_nats}
\title{Data frame for the nationalities map}
\usage{
do_map_nats(df_stats)
}
\arguments{
\item{df_stats}{Data frame with the statistics and the
corrected nationalities.}
}
\value{
List with the following elements:
\itemize{
\item df_all: Data frame with each country, its latitudes and
longitudes and whether it must be coloured or not (depending on
if there are players from that country).
\item countr_num: Vector with the countries from where there are
players and the number of them.
\item leng: Number of countries in the world.
}
}
\description{
This function prepares the data frame with the nationalities
to be mapped with \code{\link{get_map_nats}}. It is used inside it.
}
\seealso{
\code{\link{get_map_nats}}
}
\author{
Guillermo Vinue
}
