% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_offensive_fouls.R
\name{do_offensive_fouls}
\alias{do_offensive_fouls}
\title{Compute offensive fouls}
\usage{
do_offensive_fouls(data, type)
}
\arguments{
\item{data}{Play-by-play data.}

\item{type}{Either 'comm' (for committed) or 'rec' (for received).}
}
\value{
Data frame with the following columns:
\itemize{
  \strong{team}: Name of the team.
  \strong{player}: Name of the player.
  \strong{n_offensive_fouls_x}: Number of offensive fouls.
}
}
\description{
Compute how many offensive fouls has committed or received every player.
}
\examples{
df01 <- do_offensive_fouls(acb_vbc_cz_pbp_2223, "comm")
#df01

df02 <- do_offensive_fouls(acb_vbc_cz_pbp_2223, "rec")
#df02

}
\author{
Guillermo Vinue
}
