% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_preproc_period.R
\name{do_preproc_period}
\alias{do_preproc_period}
\title{Data preprocessing for periods}
\usage{
do_preproc_period(data, team_sel, period_sel, data_sl)
}
\arguments{
\item{data}{Prepared data from a given game.}

\item{team_sel}{One of the teams' names involved in the game.}

\item{period_sel}{Period of interest. Options can be "xC", where x=1,2,3,4.}

\item{data_sl}{Data with the starting lineups.}
}
\description{
Preprocess the data that will be needed for computing statistics per period.
}
\examples{
team_sel <- "Valencia Basket"
period_sel <- "1C"

pre_per <- do_preproc_period(acb_vbc_cz_pbp_2223, team_sel, period_sel, acb_vbc_cz_sl_2223)

df2 <- pre_per$df2
df0_inli_team <- pre_per$df0_inli_team

}
\author{
Guillermo Vinue
}
