% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_rpackage_stats.R
\name{do_rpackage_stats}
\alias{do_rpackage_stats}
\title{R package downloads}
\usage{
do_rpackage_stats(r_packages, year, verbose)
}
\arguments{
\item{r_packages}{Vector with the names of the R packages.}

\item{year}{String with the year.}

\item{verbose}{Should R report information on progress? TRUE or FALSE.}
}
\value{
A data frame.
}
\description{
Counts the number of times that a given R package was downloaded in a given year.
}
\examples{
\dontrun{
do_rpackage_stats(c("BAwiR", "BasketballAnalyzeR"), 2025, TRUE)
}

}
\seealso{
\link[packageRank]{cranDownloads}
}
\author{
Guillermo Vinue
}
