% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_scrape_days_acb.R
\name{do_scrape_days_acb}
\alias{do_scrape_days_acb}
\title{ACB day game codes}
\usage{
do_scrape_days_acb(edition_id)
}
\arguments{
\item{edition_id}{Identifier of the league edition. For 2026-2027 is 91. 
For coming seasons, check it at the ACB website.}
}
\value{
A data frame with two columns, one with the day and the other with 
the game code. For 34 days and 9 games per day, there must be 306 rows.
}
\description{
Obtain the game codes of any regular season day from any ACB season. 
These game codes can be used for example to define the target url 
from which collecting the shooting data of every game.
}
\note{
Before starting the web scraping, we must visit 
\url{https://www.acb.com/robots.txt} to check for permissions.
}
\examples{
\dontrun{
data_days <- do_scrape_days_acb(91)
}

}
\seealso{
\code{\link{do_scrape_shots_acb}}
}
\author{
Guillermo Vinue with help from ChatGPT.
}
