% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_violin_box_plots.R
\name{do_violin_box_plots}
\alias{do_violin_box_plots}
\title{Plots of data distributions}
\usage{
do_violin_box_plots(data_shots, data_players)
}
\arguments{
\item{data_shots}{Shooting data frame.}

\item{data_players}{Players' identifiers data frame.}
}
\value{
A plot.
}
\description{
Create violin plots and boxplots to analyze the distribution of the two-point, 
three-point and total shots. Violin plots show the distribution shape, while
boxplots give a compact statistical summary.
}
\examples{
\dontrun{
df0 <- do_divide_court_zones(acb_shooting_data_2425)

df1 <- do_filter_data(df0, "2024-2025", "", "", "", "", "")

do_violin_box_plots(df1, acb_players_2425)
}

}
\author{
Guillermo Vinue
}
