\name{eurocup_games_1718}
\alias{eurocup_games_1718}
\docType{data}
\title{
Eurocup games 2017-2018
}
\description{
Games of the ten days of regular season and the first three days of top 16 
of the Eurocup 2017-2018 season.
}
\usage{eurocup_games_1718}
\format{Data frame with 3604 rows and 38 columns.}
\source{\url{https://www.eurocupbasketball.com/}}
\keyword{datasets}
