\name{eurocup_players_1718}
\alias{eurocup_players_1718}
\docType{data}
\title{
Eurocup players 2017-2018
}
\description{
Players corresponding to the games of the ten days of regular season 
and the first three days of top 16 of the Eurocup 2017-2018 season.
}
\usage{eurocup_players_1718}
\format{Data frame with 351 rows and 7 columns.}
\source{\url{http://www.eurocupbasketball.com/}}
\keyword{datasets}
