\name{euroleague_games_1718}
\alias{euroleague_games_1718}
\docType{data}
\title{
Euroleague games 2017-2018
}
\description{
Games of the first nineteen days of the Euroleague 2017-2018 season.
}
\usage{euroleague_games_1718}
\format{Data frame with 3932 rows and 38 columns.}
\source{\url{http://www.euroleague.net}}
\keyword{datasets}
