\name{euroleague_players_1718}
\alias{euroleague_players_1718}
\docType{data}
\title{
Euroleague players 2017-2018
}
\description{
Players corresponding to the games of the first nineteen days of the 
Euroleague 2017-2018 season.
}
\usage{euroleague_players_1718}
\format{Data frame with 245 rows and 7 columns.}
\source{\url{https://www.euroleaguebasketball.net/euroleague/}}
\keyword{datasets}
