% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_donut_usage.R
\name{get_donut_usage}
\alias{get_donut_usage}
\title{Donut chart usage possessions}
\usage{
get_donut_usage(data_usage, team_sel, size_play, size_perc)
}
\arguments{
\item{data_usage}{Data frame with the number of possessions that each 
player played and the number that he ended.}

\item{team_sel}{String with the team's full name.}

\item{size_play}{Size of the players' labels.}

\item{size_perc}{Size of the percentages labels.}
}
\value{
A donut chart.
}
\description{
This function creates a donut chart that displays the percentage of 
possessions that each player of a team ends while being on the court.
}
\details{
Donut charts are an alternative for pie charts, which have a hole 
in the middle, making them cleaner to read than pie charts.
}
\examples{
\dontrun{
get_donut_usage(acb_usage_data_2526, "Valencia Basket", 3, 4)
}

}
\seealso{
\code{\link{get_donut_usage_action}}
}
\author{
Guillermo Vinue with the help of ChatGPT.
}
