% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_pyramid_acb.R
\name{get_pop_pyramid_acb}
\alias{get_pop_pyramid_acb}
\title{ACB population pyramid}
\usage{
get_pop_pyramid_acb(df)
}
\arguments{
\item{df}{Data frame that contains the ACB players' nationality.}
}
\value{
Graphical device
}
\description{
This is the code to get a population pyramid with the number of both Spanish 
and foreigner players along the ACB seasons. This aids in discussion of nationality 
imbalance.
}
\examples{
\dontrun{
 # Load the data_app_acb file with the ACB games 
 # from seasons 1985-1986 to 2017-2018:
 load(url("http://www.uv.es/vivigui/softw/data_app_acb.RData"))
 get_pop_pyramid_acb(data_app_acb)
}

}
\author{
Guillermo Vinue
}
