% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roster_age_profile.R
\name{get_roster_age_profile}
\alias{get_roster_age_profile}
\title{Roster age profile}
\usage{
get_roster_age_profile(data_age_team, team_sel, language)
}
\arguments{
\item{data_age_team}{Data frame with the team's age profile.}

\item{team_sel}{Team.}

\item{language}{Language labels. Current options are 'en' for English
and 'es' for Spanish.}
}
\value{
Graphical device.
}
\description{
For the players of the same team, show their age at time of joining the team,
their current year and how many years they have spent in the team.
}
\examples{
\dontrun{
get_roster_age_profile(acb_age_profile_data_2526, "Valencia Basket", "es")
}

}
\author{
Guillermo Vinue
}
