\name{acb_combs}
\alias{acb_combs}
\docType{data}
\title{
Lineups in a given game.
}
\description{
Data set for illustrative purposes with the Unicaja five-men and one-man lineups 
in the game 104459 from the ACB 2025-2026 regular season.
}
\usage{acb_combs}
\format{Data frame with 150 rows and 38 columns.}
\source{\url{https://www.acb.com/}}
\keyword{datasets}
