\name{acb_games_2223_coach}
\alias{acb_games_2223_coach}
\docType{data}
\title{
ACB coaches in the 2022-2023 season.
}
\description{
Coach for each team in all the games of the ACB 2022-2023 season.
}
\note{
The \strong{game_code} column allows us to detect the source website, for example,
\url{https://live.acb.com/es/partidos/103389/jugadas}.
}
\usage{acb_games_2223_coach}
\format{Data frame with 612 rows and 4 columns.}
\source{\url{https://www.acb.com/}}
\keyword{datasets}
