\name{acb_games_2223_info}
\alias{acb_games_2223_info}
\docType{data}
\title{
ACB games 2022-2023, days and codes.
}
\description{
Game codes, games and days from the ACB 2022-2023 season.
}
\note{
The \strong{game_code} column allows us to detect the source website, for example,
\url{https://live.acb.com/es/partidos/103389/jugadas}.
}
\usage{acb_games_2223_info}
\format{Data frame with 306 rows and 3 columns.}
\source{\url{https://www.acb.com/}}
\keyword{datasets}
