\name{acb_players_1718}
\alias{acb_players_1718}
\docType{data}
\title{
ACB players 2017-2018
}
\description{
Players corresponding to the games of the first seventeen 
days of the ACB 2017-2018 season.
}
\usage{acb_players_1718}
\format{Data frame with 255 rows and 7 columns.}
\source{\url{https://www.acb.com/}}
\keyword{datasets}
