\name{acb_players_2425}
\alias{acb_players_2425}
\docType{data}
\title{
ACB players 2024-2025
}
\description{
Player unique identifiers of a sample of 30 players from the ACB 2024-2025 season.
}
\usage{acb_players_2425}
\format{Data frame with 30 rows and 5 columns.}
\source{\url{https://www.acb.com/}}
\keyword{datasets}
