\name{acb_shooting_data_2425}
\alias{acb_shooting_data_2425}
\docType{data}
\title{
ACB shooting data, 2024-2025
}
\description{
Spatial shooting data from a sample of 30 players from the ACB 2024-2025 season.
See also \code{acb_players_2425}.
}
\usage{acb_shooting_data_2425}
\format{Data frame with 4277 rows and 23 columns.}
\source{\url{https://www.acb.com/}}
\keyword{datasets}
