\name{acb_vbc_cz_pbp_2223}
\alias{acb_vbc_cz_pbp_2223}
\docType{data}
\title{
ACB play-by-play data, 2022-2023, Valencia Basket-Casademont Zaragoza
}
\description{
Play-by-play data from the game Valencia Basket-Casademont Zaragoza 
from the ACB 2022-2023 season. 
}
\note{
Actions are given in Spanish. A bilingual basketball vocabulary (Spanish/English)
is provided in \url{https://www.uv.es/vivigui/docs/basketball_dictionary.xlsx}.
The \strong{game_code} column allows us to detect the source website, namely,
\url{https://jv.acb.com/es/103389/jugadas}.
}
\usage{acb_vbc_cz_pbp_2223}
\format{Data frame with 466 rows and 9 columns.}
\source{\url{https://www.acb.com/}}
\keyword{datasets}
