\name{acb_vbc_cz_sl_2223}
\alias{acb_vbc_cz_sl_2223}
\docType{data}
\title{
ACB starting lineups, 2022-2023, Valencia Basket-Casademont Zaragoza
}
\description{
Starting lineups in each period from the game Valencia Basket-Casademont Zaragoza 
from the ACB 2022-2023 season.
}
\note{
The \strong{action} column refers to starting lineup (\emph{Quinteto inicial}, in Spanish).
The initial score in each period does not really matter for the creation of this data set.
The \strong{game_code} column allows us to detect the source website, for example,
\url{https://jv.acb.com/es/103389/jugadas}.
}
\usage{acb_vbc_cz_sl_2223}
\format{Data frame with 40 rows and 9 columns.}
\source{\url{https://www.acb.com/}}
\keyword{datasets}
