% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_clutch_time.R
\name{do_clutch_time}
\alias{do_clutch_time}
\title{Get games with clutch time}
\usage{
do_clutch_time(data)
}
\arguments{
\item{data}{Source play-by-play data.}
}
\value{
Data frame of the game that has clutch time.
}
\description{
Obtain the games that have clutch time. The clutch time is the game situation when the 
scoring margin is within 5 points with five or fewer minutes remaining in a game.
}
\examples{
df0 <- do_clutch_time(acb_vbc_cz_pbp_2223)
#df0 # If no rows, that means that the game did not have clutch time.

}
\author{
Guillermo Vinue
}
