\name{metrics_player_zone}
\alias{metrics_player_zone}
\docType{data}
\title{
ACB players 2024-2025
}
\description{
Metrics (total shots, field goal percentages and points per shot) at some of the 
court zones for some of the 30 players contained in the data frame \code{acb_players_2425}.
}
\usage{metrics_player_zone}
\format{Data frame with 8 rows and 5 columns.}
\source{\url{https://www.acb.com/}}
\keyword{datasets}
